+++ core/src/codegen/trait_impl.rs
@@ -1,10 +1,11 @@
 use quote::Tokens;
 use syn::{Generics, Ident, Path, WherePredicate};
 
-use ast::Data;
+use ast::{Data, Fields};
 use codegen::error::{ErrorCheck, ErrorDeclaration};
 use codegen::field;
 use codegen::{DefaultExpression, Field, FieldsGen, Variant};
+use usage::{CollectTypeParams, IdentSet, Purpose};
 
 #[derive(Debug)]
 pub struct TraitImpl<'a> {
@@ -14,6 +15,69 @@ pub struct TraitImpl<'a> {
     pub default: Option<DefaultExpression<'a>>,
     pub map: Option<&'a Path>,
     pub bound: Option<&'a [WherePredicate]>,
+}
+
+impl<'a> TraitImpl<'a> {
+    /// Get all declared type parameters.
+    pub fn declared_type_params(&self) -> IdentSet {
+        self.generics
+            .type_params()
+            .map(|tp| tp.ident.clone())
+            .collect()
+    }
+
+    /// Get the type parameters which are used by non-skipped fields.
+    pub fn used_type_params(&self) -> IdentSet {
+        self.type_params_matching(|f| !f.skip, |v| !v.skip)
+    }
+
+    /// Get the type parameters which are used by skipped fields.
+    pub fn skipped_type_params(&self) -> IdentSet {
+        self.type_params_matching(|f| f.skip, |v| v.skip)
+    }
+
+    fn type_params_matching<'b, F, V>(&'b self, field_filter: F, variant_filter: V) -> IdentSet
+    where
+        F: Fn(&&Field) -> bool,
+        V: Fn(&&Variant) -> bool,
+    {
+        let declared = self.declared_type_params();
+        match self.data {
+            Data::Struct(ref v) => self.type_params_in_fields(v, &field_filter, &declared),
+            Data::Enum(ref v) => v.iter().filter(variant_filter).fold(
+                Default::default(),
+                |mut state, variant| {
+                    state.extend(self.type_params_in_fields(
+                        &variant.data,
+                        &field_filter,
+                        &declared,
+                    ));
+                    state
+                },
+            ),
+        }
+    }
+
+    /// Get the type parameters of all fields in a set matching some filter
+    fn type_params_in_fields<'b, F>(
+        &'b self,
+        fields: &'b Fields<Field<'a>>,
+        field_filter: F,
+        declared: &IdentSet,
+    ) -> IdentSet
+    where
+        F: Fn(&&'b Field) -> bool,
+    {
+        let mut hits = IdentSet::default();
+        hits.extend(
+            fields
+                .iter()
+                .filter(field_filter)
+                .collect_type_params(&Purpose::BoundImpl.into(), declared),
+        );
+
+        hits
+    }
 }
 
 impl<'a> TraitImpl<'a> {