+++ src/macros_public.rs
@@ -1,0 +1,1 @@
+../core/src/macros_public.rs